+++ src/storage/models.rs
@@ -26,7 +26,7 @@ use tablenames::*;
 pub struct AssetDesc {
     pub symbol: String,
     pub name: String,
-    pub chain_id: u32,
+    pub chain_id: i16,
     pub token_address: String,
     pub is_commonly_quoted: bool,
     pub precision_stor: i16,
